Migration of data from LiteDB to shared pref storage

+++ src/App/Services/MobileStorageService.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 using Bit.Core;
 using Bit.Core.Abstractions;
@@ -11,28 +12,23 @@ namespace Bit.App.Services
         private readonly IStorageService _preferencesStorageService;
         private readonly IStorageService _liteDbStorageService;
 
-        private readonly HashSet<string> _preferenceStorageKeys = new HashSet<string>
+        private readonly HashSet<string> _liteDbStorageKeys = new HashSet<string>
         {
-            Constants.StateVersionKey,
-            Constants.PreAuthEnvironmentUrlsKey,
-            Constants.AutofillTileAdded,
-            Constants.AddSitePromptShownKey,
-            Constants.PushInitialPromptShownKey,
-            Constants.LastFileCacheClearKey,
-            Constants.PushRegisteredTokenKey,
-            Constants.LastBuildKey,
-            Constants.ClearCiphersCacheKey,
-            Constants.BiometricIntegritySourceKey,
-            Constants.iOSExtensionActiveUserIdKey,
-            Constants.iOSAutoFillClearCiphersCacheKey,
-            Constants.iOSAutoFillBiometricIntegritySourceKey,
-            Constants.iOSExtensionClearCiphersCacheKey,
-            Constants.iOSExtensionBiometricIntegritySourceKey,
-            Constants.iOSShareExtensionClearCiphersCacheKey,
-            Constants.iOSShareExtensionBiometricIntegritySourceKey,
-            Constants.RememberedEmailKey,
-            Constants.RememberedOrgIdentifierKey
+            Constants.EventCollectionKey,
+            Constants.CiphersKey(""),
+            Constants.FoldersKey(""),
+            Constants.CollectionsKey(""),
+            Constants.CiphersLocalDataKey(""),
+            Constants.SendsKey(""),
+            Constants.PassGenHistoryKey(""),
+            Constants.SettingsKey(""),
         };
+
+        private bool IsLiteDbKey(string key)
+        {
+            return _liteDbStorageKeys.Any(key.StartsWith) ||
+                   _liteDbStorageKeys.Contains(key);
+        }
 
         public MobileStorageService(
             IStorageService preferenceStorageService,
@@ -44,29 +40,42 @@ namespace Bit.App.Services
 
         public async Task<T> GetAsync<T>(string key)
         {
-            if (_preferenceStorageKeys.Contains(key))
+            if (IsLiteDbKey(key))
             {
-                return await _preferencesStorageService.GetAsync<T>(key);
+                return await _liteDbStorageService.GetAsync<T>(key);
             }
-            return await _liteDbStorageService.GetAsync<T>(key);
+
+            var result = await _preferencesStorageService.GetAsync<T>(key);
+            if (result == null)
+            {
+                // If result is expected in prefs but found in LiteDB, migrate it to prefs
+                result = await _liteDbStorageService.GetAsync<T>(key);
+                if (result != null)
+                {
+                    await _preferencesStorageService.SaveAsync(key, result);
+                    await _liteDbStorageService.RemoveAsync(key);
+                }
+            }
+
+            return result;
         }
 
         public Task SaveAsync<T>(string key, T obj)
         {
-            if (_preferenceStorageKeys.Contains(key))
+            if (IsLiteDbKey(key))
             {
-                return _preferencesStorageService.SaveAsync(key, obj);
+                return _liteDbStorageService.SaveAsync(key, obj);
             }
-            return _liteDbStorageService.SaveAsync(key, obj);
+            return _preferencesStorageService.SaveAsync(key, obj);
         }
 
         public Task RemoveAsync(string key)
         {
-            if (_preferenceStorageKeys.Contains(key))
+            if (IsLiteDbKey(key))
             {
-                return _preferencesStorageService.RemoveAsync(key);
+                return _liteDbStorageService.RemoveAsync(key);
             }
-            return _liteDbStorageService.RemoveAsync(key);
+            return _preferencesStorageService.RemoveAsync(key);
         }
 
         public void Dispose()
